Replace use of "whitelist" with "allowlist"
This change replaces our use of the term "whitelist" with "allowlist" where we can.

"Allowlist" and "blocklist" have a clarity of meaning for which "whitelist" and "blacklist" rely on metaphor, and metaphor can encode in thought and language both historic and current forms of domination. This is an attempt to rid our codebase of two of those metaphors.